+++ lib/Models/Cesium.ts
@@ -65,9 +65,10 @@ import TileErrorHandlerMixin from "../ModelMixins/TileErrorHandlerMixin";
 import SplitterTraits from "../Traits/TraitsClasses/SplitterTraits";
 import TerriaViewer from "../ViewModels/TerriaViewer";
 import CameraView from "./CameraView";
-import hasTraits from "./Definition/hasTraits";
 import Feature from "./Feature";
 import GlobeOrMap from "./GlobeOrMap";
+import hasTraits from "./Definition/hasTraits";
+import MapInteractionMode from "./MapInteractionMode";
 import Terria from "./Terria";
 import UserDrawing from "./UserDrawing";
 
@@ -303,8 +304,7 @@ export default class Cesium extends GlobeOrMap {
 
     // Handle left click by picking objects from the map.
     inputHandler.setInputAction(e => {
-      if (!this.isFeaturePickingPaused)
-        this.pickFromScreenPosition(e.position, false);
+      if (!this.isFeaturePickingPaused) this.pickFromScreenPosition(e.position);
     }, ScreenSpaceEventType.LEFT_CLICK);
 
     let zoomUserDrawing: UserDrawing | undefined;
@@ -992,7 +992,7 @@ export default class Cesium extends GlobeOrMap {
    * specified and set terria.pickedFeatures based on this.
    *
    */
-  pickFromScreenPosition(screenPosition: Cartesian2, ignoreSplitter: boolean) {
+  pickFromScreenPosition(screenPosition: Cartesian2, ignoreSplitter?: boolean) {
     const pickRay = this.scene.camera.getPickRay(screenPosition);
     const pickPosition = this.scene.globe.pick(pickRay, this.scene);
     const pickPositionCartographic =
@@ -1005,6 +1005,10 @@ export default class Cesium extends GlobeOrMap {
       ? this.scene.imageryLayers.pickImageryLayerFeatures(pickRay, this.scene)
       : undefined;
 
+    const mapInteractionMode: MapInteractionMode | undefined = this.terria
+      .mapInteractionModeStack[this.terria.mapInteractionModeStack.length - 1];
+    const shouldIgnoreSplitter =
+      ignoreSplitter ?? mapInteractionMode?.ignoreSplitterWhenPicking;
     const result = this._buildPickedFeatures(
       providerCoords,
       pickPosition,
@@ -1012,18 +1016,12 @@ export default class Cesium extends GlobeOrMap {
       pickRasterPromise ? [pickRasterPromise] : [],
       undefined,
       pickPositionCartographic ? pickPositionCartographic.height : 0.0,
-      ignoreSplitter
+      shouldIgnoreSplitter ?? false
     );
 
-    const mapInteractionModeStack = this.terria.mapInteractionModeStack;
     runInAction(() => {
-      if (
-        isDefined(mapInteractionModeStack) &&
-        mapInteractionModeStack.length > 0
-      ) {
-        mapInteractionModeStack[
-          mapInteractionModeStack.length - 1
-        ].pickedFeatures = result;
+      if (mapInteractionMode) {
+        mapInteractionMode.pickedFeatures = result;
       } else {
         this.terria.pickedFeatures = result;
       }
@@ -1074,6 +1072,9 @@ export default class Cesium extends GlobeOrMap {
       }
     }
 
+    const mapInteractionMode: MapInteractionMode | undefined = this.terria
+      .mapInteractionModeStack[this.terria.mapInteractionModeStack.length - 1];
+
     const result = this._buildPickedFeatures(
       providerCoords,
       pickPosition,
@@ -1081,17 +1082,11 @@ export default class Cesium extends GlobeOrMap {
       filterOutUndefined(promises),
       imageryLayers,
       pickPositionCartographic.height,
-      false
+      mapInteractionMode?.ignoreSplitterWhenPicking
     );
 
-    const mapInteractionModeStack = this.terria.mapInteractionModeStack;
-    if (
-      defined(mapInteractionModeStack) &&
-      mapInteractionModeStack.length > 0
-    ) {
-      mapInteractionModeStack[
-        mapInteractionModeStack.length - 1
-      ].pickedFeatures = result;
+    if (mapInteractionMode) {
+      mapInteractionMode.pickedFeatures = result;
     } else {
       this.terria.pickedFeatures = result;
     }