+++ lib/Models/Terria.ts
@@ -67,6 +67,7 @@ import MagdaReference, {
   MagdaReferenceHeaders
 } from "./Catalog/CatalogReferences/MagdaReference";
 import SplitItemReference from "./Catalog/CatalogReferences/SplitItemReference";
+import { CompareConfig, createCompareConfig } from "./Comparable";
 import CommonStrata from "./Definition/CommonStrata";
 import hasTraits from "./Definition/hasTraits";
 import { BaseModel } from "./Definition/Model";
@@ -503,6 +504,9 @@ export default class Terria {
   @observable terrainSplitDirection: ImagerySplitDirection =
     ImagerySplitDirection.NONE;
 
+  // Is true when the workflow panel is active
+  @observable isWorkflowPanelActive = false;
+
   @observable depthTestAgainstTerrainEnabled = false;
 
   @observable stories: any[] = [];
@@ -541,6 +545,11 @@ export default class Terria {
    * @type {boolean}
    */
   @observable catalogReferencesLoaded: boolean = false;
+
+  /**
+   * Configuration for the compare workflow
+   */
+  @observable compareConfig: CompareConfig | undefined = undefined;
 
   augmentedVirtuality?: any;
 
@@ -1399,6 +1408,12 @@ export default class Terria {
 
     if (isJsonNumber(initData.splitPosition)) {
       this.splitPosition = initData.splitPosition;
+    }
+
+    this.isWorkflowPanelActive = Boolean(initData.isWorkflowPanelActive);
+
+    if (isJsonObject(initData.compareConfig)) {
+      this.compareConfig = createCompareConfig(initData.compareConfig);
     }
 
     // Copy but don't yet load the workbench.