+++ lib/ReactViews/Workbench/Controls/Legend.tsx
@@ -6,14 +6,14 @@ import defined from "terriajs-cesium/Source/Core/defined";
 import Resource from "terriajs-cesium/Source/Core/Resource";
 import URI from "urijs";
 import isDefined from "../../../Core/isDefined";
-import CatalogMemberMixin from "../../../ModelMixins/CatalogMemberMixin";
-import Model from "../../../Models/Definition/Model";
+import MinMaxLevelMixin from "../../../ModelMixins/MinMaxLevelMixin";
 import proxyCatalogItemUrl from "../../../Models/Catalog/proxyCatalogItemUrl";
+import Model, { BaseModel } from "../../../Models/Definition/Model";
 import LegendTraits, {
   LegendItemTraits
 } from "../../../Traits/TraitsClasses/LegendTraits";
 import Styles from "./legend.scss";
-import MinMaxLevelMixin from "../../../ModelMixins/MinMaxLevelMixin";
+import CatalogMemberMixin from "../../../ModelMixins/CatalogMemberMixin";
 
 /* A lookup map for displayable mime types */
 const DISPLAYABLE_MIME_TYPES = [
@@ -38,7 +38,7 @@ function checkMimeType(legend: Model<LegendTraits>) {
 
 @observer
 export default class Legend extends React.Component<{
-  item: CatalogMemberMixin.Instance;
+  item: BaseModel;
   forPrint?: boolean;
 }> {
   static defaultProps = {
@@ -252,6 +252,7 @@ export default class Legend extends React.Component<{
 
   render() {
     if (
+      !CatalogMemberMixin.isMixedInto(this.props.item) ||
       this.props.item.hideLegendInWorkbench ||
       (MinMaxLevelMixin.isMixedInto(this.props.item) &&
         this.props.item.scaleWorkbenchInfo)