+++ lib/ReactViews/Workbench/Controls/OpacitySection.jsx
@@ -1,5 +1,3 @@
-"use strict";
-
 import createReactClass from "create-react-class";
 import { runInAction } from "mobx";
 import { observer } from "mobx-react";
@@ -59,4 +57,4 @@ const OpacitySection = observer(
   })
 );
 
-module.exports = withTranslation()(OpacitySection);
+export default withTranslation()(OpacitySection);